requirements.txt
google-generativeai
IPython
Pillow
aw-client
pytz
